+++ src/Adapter/AmqpLib/Queue/AmqpQueueFactory.php
@@ -30,6 +30,11 @@ class AmqpQueueFactory implements QueueFactoryInterface, \SplObserver
     ) {
     }
 
+    public function withPassive(): QueueFactoryInterface
+    {
+        return new AmqpQueueFactory($this->channelFactory, $this->definition->withPassive(true));
+    }
+
     public function create(): QueueInterface
     {
         if ($this->queue) {
@@ -45,12 +50,16 @@ class AmqpQueueFactory implements QueueFactoryInterface, \SplObserver
             ));
         }
 
-        /** @var AmqpConnection $connection */
         $connection = $channel->getConnection();
-        $connection->attach($this);
 
         $queue = new AmqpQueue($channel, $this->definition);
         $queue->declare();
+
+        if ($this->definition->passive) {
+            return $queue;
+        }
+
+        $connection->attach($this);
 
         foreach ($this->definition->bindings as $binding) {
             $this->bind($binding->exchangeName, $binding->routingKey);